Refactored some code in different files
