+++ Viewer2D.cpp
@@ -15,6 +15,8 @@
 #include "rapidjson/writer.h"
 #include <string>
 #include <memory>
+#include "GridLayer.h"
+#include "SceneLayer.h"
 
 
 void GLError(std::string name)
@@ -43,8 +45,8 @@ Viewer2D::Viewer2D(QWidget *parent)
     : QOpenGLWidget(parent), layers()
 {
     std::cout << "create layers" << std::endl;
+    this->layers.push_back(std::shared_ptr<Layer>(new SceneLayer()));
     this->layers.push_back(std::shared_ptr<Layer>(new GridLayer()));
-    //this->layers.push_back(std::shared_ptr<Layer>(new SceneLayer()));
 }
 
 void Viewer2D::SetOperationRunner(std::shared_ptr<OperationRunner> operationRunner)
@@ -104,76 +106,7 @@ void Viewer2D::UpdateFromModel(std::shared_ptr<Model> const &model)
     }
 }
 
-GridLayer::GridLayer(): positionsBuffer(QOpenGLBuffer::VertexBuffer)
-{
 
-}
-
-void GridLayer::InitializeGL(QObject* context, std::unique_ptr<QOpenGLFunctions, void(*)(void*)> const &f)
-{
-    QColor color(255, 255, 255, 255);
-
-    gridSpacing = 0.2f;
-
-    positionsBuffer.setUsagePattern(QOpenGLBuffer::DynamicDraw);
-    positionsBuffer.create();
-
-    std::unique_ptr<std::string> vertexFile = std::unique_ptr<std::string>(new std::string("GPU\\Grid.vert.glsl"));
-    std::unique_ptr<std::string> fragmentFile = std::unique_ptr<std::string>(new std::string("GPU\\Grid.frag.glsl"));
-
-    program = std::unique_ptr<QOpenGLShaderProgram>(new QOpenGLShaderProgram(nullptr));
-    program->addShaderFromSourceFile(QOpenGLShader::Vertex, QString::fromStdString(*vertexFile));
-    program->addShaderFromSourceFile(QOpenGLShader::Fragment, QString::fromStdString(*fragmentFile));
-    program->link();
-
-    program->bind();
-
-    program->enableAttributeArray("a_position");
-    program->setUniformValue("u_color", color);
-}
-
-void GridLayer::PaintGL(QObject* context, std::unique_ptr<QOpenGLFunctions, void(*)(void*)> const &f)
-{
-    program->bind();
-    f->glLineWidth(1.0f);
-    f->glDrawArrays(GL_LINES, 0, lines*2);
-}
-
-void GridLayer::UpdateScene(std::shared_ptr<Model> const &m, std::unique_ptr<QOpenGLFunctions, void(*)(void*)> const &f)
-{
-    this->gridSpacing = (*m->d->GetSceneConf())["grid_spacing"].GetDouble();
-    if(m->view->IsChanged())
-    {
-        UpdateLines();
-
-        positionsBuffer.bind();
-        f->glVertexAttribPointer((GLuint)program->attributeLocation("a_position"), 2, GL_FLOAT, GL_FALSE, 0, NULL);
-        positionsBuffer.allocate(this->positions->data(), this->positions->size()*sizeof(float));
-
-        program->bind();
-        program->setUniformValue("u_view", this->viewMatrix);
-    }
-}
-
-MinMaxValue GridLayer::GetMinMax()
-{
-    MinMaxValue result;
-    result.Active = false;
-    return result;
-}
-
-float GridLayer::CalcIdealSpacing()
-{
-    QVector3D center = this->viewMatrix * QVector3D(0, 0, 0);
-    QVector3D gridSpacingPoint = this->viewMatrix * QVector3D(this->gridSpacing, 0, 0);
-    float distance = (center - gridSpacingPoint).length();
-    float x = 1;
-    while(0.04 > distance * x)
-    {
-        x *= 2;
-    }
-    return this->gridSpacing * x;
-}
 
 
 void Viewer2D::UpdateViewMatrix(QMatrix4x4 matrix)
@@ -181,50 +114,6 @@ void Viewer2D::UpdateViewMatrix(QMatrix4x4 matrix)
     this->_view_matrix = matrix;
 
     std::for_each(this->layers.begin(), this->layers.end(), [matrix](std::shared_ptr<Layer> l){l->UpdateViewMatrix(matrix);});
-}
-
-void GridLayer::UpdateLines()
-{
-    float grid_spacing = CalcIdealSpacing();
-
-    QVector2D tl = (this->viewMatrix.inverted() * QVector3D(-1, -1, 0)).toVector2D();
-    QVector2D br = (this->viewMatrix.inverted() * QVector3D(1, 1, 0)).toVector2D();
-
-    tl[0] = floorf(tl[0]/grid_spacing);
-    tl[1] = floorf(tl[1]/grid_spacing);
-    br[0] = ceilf(br[0]/grid_spacing);
-    br[1] = ceilf(br[1]/grid_spacing);
-
-    unsigned int verticalLines = (unsigned int)fabsf(br[0]-tl[0]);
-    unsigned int horizontalLines = (unsigned int)fabsf(br[1]-tl[1]);
-
-    unsigned int lines = verticalLines+horizontalLines;
-
-    tl[0] = tl[0]*grid_spacing;
-    tl[1] = tl[1]*grid_spacing;
-    br[0] = br[0]*grid_spacing;
-    br[1] = br[1]*grid_spacing;
-
-    std::unique_ptr<std::vector<float> > positions = std::unique_ptr<std::vector<float> >(new std::vector<float>(lines*4));
-    for(int i = 0; i < verticalLines; i++)
-    {
-        (*positions)[i*4+0] = tl[0]+i*grid_spacing;
-        (*positions)[i*4+1] = tl[1];
-        (*positions)[i*4+2] = tl[0]+i*grid_spacing;
-        (*positions)[i*4+3] = br[1];
-    }
-
-    for(int i = 0; i < horizontalLines; i++)
-    {
-        int offset = verticalLines*4;
-        (*positions)[offset+i*4+0] = tl[0];
-        (*positions)[offset+i*4+1] = tl[1]+i*grid_spacing;
-        (*positions)[offset+i*4+2] = br[0];
-        (*positions)[offset+i*4+3] = tl[1]+i*grid_spacing;
-    }
-
-    this->positions = std::move(positions);
-    this->lines = lines;
 }
 
 MinMaxValue::MinMaxValue()
@@ -284,152 +173,6 @@ MinMaxValue MinMaxValue::CombineList(std::vector<MinMaxValue> list)
     }
     return result;
 }
-
-SceneLayer::SceneLayer() : positions(new std::vector<float>()), values(new std::vector<float>()), lines(0),
-               positionsBuffer(QOpenGLBuffer::VertexBuffer), valuesBuffer(QOpenGLBuffer::VertexBuffer)
-{
-
-}
-
-void SceneLayer::InitializeGL(QObject *context, std::unique_ptr<QOpenGLFunctions, void (*)(void *)> const &f)
-{
-    positionsBuffer.setUsagePattern(QOpenGLBuffer::DynamicDraw);
-    positionsBuffer.create();
-
-    valuesBuffer.setUsagePattern(QOpenGLBuffer::DynamicDraw);
-    valuesBuffer.create();
-
-    std::unique_ptr<std::string> vertexFile = std::unique_ptr<std::string>(new std::string("GPU\\Scene2D.vert"));
-    std::unique_ptr<std::string> fragmentFile = std::unique_ptr<std::string>(new std::string("GPU\\Scene2D.frag"));
-
-    program = std::unique_ptr<QOpenGLShaderProgram>(new QOpenGLShaderProgram(nullptr));
-    program->addShaderFromSourceFile(QOpenGLShader::Vertex, QString::fromStdString(*vertexFile));
-    program->addShaderFromSourceFile(QOpenGLShader::Fragment, QString::fromStdString(*fragmentFile));
-    program->link();
-
-    program->bind();
-
-    CreateColormap();
-
-    program->enableAttributeArray("a_position");
-    program->enableAttributeArray("a_value");
-    program->setUniformValue("colormap", this->textureID);
-    program->setUniformValue("vmin", 0.0f);
-    program->setUniformValue("vmax", 1.0f);
-
-
-}
-
-void SceneLayer::PaintGL(QObject *context, std::unique_ptr<QOpenGLFunctions, void (*)(void *)> const &f)
-{
-    program->bind();
-    f->glLineWidth(5.0f);
-    f->glDrawArrays(GL_LINES, 0, lines*2);
-}
-
-void SceneLayer::UpdateScene(std::shared_ptr<Model> const &m, std::unique_ptr<QOpenGLFunctions, void(*)(void*)> const &f)
-{
-    if(m->view->IsChanged())
-    {
-        program->bind();
-        program->setUniformValue("u_view", this->viewMatrix);
-    }
-
-    std::shared_ptr<rapidjson::Document> conf = m->d->GetSceneConf();
-
-    rapidjson::Value& domains = (*conf)["domains"];
-
-    this->positions->clear();
-    this->values->clear();
-    this->lines = 0;
-
-    for(rapidjson::SizeType i = 0; i < domains.Size(); i++)
-    {
-        QVector2D tl(domains[i]["topleft"][0].GetDouble(), domains[i]["topleft"][1].GetDouble());
-        QVector2D size(domains[i]["size"][0].GetDouble(), domains[i]["size"][1].GetDouble());
-        QVector2D br = tl+size;
-        float left = tl[0];
-        float top = tl[1];
-
-        float right = br[0];
-        float bottom = br[1];
-
-        float aTop = domains[i]["edges"]["t"]["a"].GetDouble();
-        float aBottom = domains[i]["edges"]["b"]["a"].GetDouble();
-        float aLeft = domains[i]["edges"]["l"]["a"].GetDouble();
-        float aRight = domains[i]["edges"]["r"]["a"].GetDouble();
-
-        this->positions->push_back(left); this->positions->push_back(top);
-        this->values->push_back(aTop);
-        this->positions->push_back(right); this->positions->push_back(top);
-        this->values->push_back(aTop);
-
-        this->positions->push_back(left); this->positions->push_back(top);
-        this->values->push_back(aLeft);
-        this->positions->push_back(left); this->positions->push_back(bottom);
-        this->values->push_back(aLeft);
-
-        this->positions->push_back(left); this->positions->push_back(bottom);
-        this->values->push_back(aBottom);
-        this->positions->push_back(right); this->positions->push_back(bottom);
-        this->values->push_back(aBottom);
-
-        this->positions->push_back(right); this->positions->push_back(top);
-        this->values->push_back(aRight);
-        this->positions->push_back(right); this->positions->push_back(bottom);
-        this->values->push_back(aRight);
-
-        this->lines += 4;
-    }
-
-    positionsBuffer.bind();
-    f->glVertexAttribPointer((GLuint)program->attributeLocation("a_position"), 2, GL_FLOAT, GL_FALSE, 0, 0);
-    positionsBuffer.allocate(this->positions->data(), this->positions->size()*sizeof(float));
-
-    valuesBuffer.bind();
-    f->glVertexAttribPointer((GLuint)program->attributeLocation("a_value"), 1, GL_FLOAT, GL_FALSE, 0, 0);
-    valuesBuffer.allocate(this->values->data(), this->values->size()*sizeof(float));
-}
-
-MinMaxValue SceneLayer::GetMinMax()
-{
-    return MinMaxValue();
-}
-
-
-
-void SceneLayer::CreateColormap()
-{
-    std::unique_ptr<std::vector<float>> colormap(new std::vector<float>(2*512*4));
-    for(int i = 0; i < 512; i++)
-    {
-        (*colormap)[i*4+0] = i/512.0f;
-        (*colormap)[i*4+1] = 0;
-        (*colormap)[i*4+2] = 1-i/512.0f;
-        (*colormap)[i*4+3] = 1;
-
-        (*colormap)[512*4+i*4+0] = i/512.0f;
-        (*colormap)[512*4+i*4+1] = 0;
-        (*colormap)[512*4+i*4+2] = 1-i/512.0f;
-        (*colormap)[512*4+i*4+3] = 1;
-    }
-
-    // Create one OpenGL texture
-    GLuint textureID;
-    glGenTextures(1, &textureID);
-
-    // "Bind" the newly created texture : all future texture functions will modify this texture
-    glBindTexture(GL_TEXTURE_2D, textureID);
-
-    // Give the image to OpenGL
-    glTexImage2D(GL_TEXTURE_2D, 0, GL_RGBA32F, 512, 2, 0, GL_RGBA, GL_FLOAT, colormap->data());
-
-    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);
-    glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
-
-    this->textureID = textureID;
-}
-
 
 void Viewer2D::mousePressEvent(QMouseEvent *event)
 {
