+++ Viewer2D.h
@@ -83,54 +83,4 @@ public:
     virtual void UpdateViewMatrix(QMatrix4x4 viewMatrix){ this->viewMatrix = viewMatrix; };
 };
 
-class GridLayer: public Layer
-{
-private:
-    std::unique_ptr<QOpenGLShaderProgram> program;
-    void UpdateLines();
-    std::unique_ptr<std::vector<float>> positions;
-    QOpenGLBuffer positionsBuffer;
-    int lines;
-    float gridSpacing;
-
-public:
-    GridLayer();
-
-    virtual void InitializeGL(QObject* context, std::unique_ptr<QOpenGLFunctions, void(*)(void*)> const &f);
-
-    virtual void PaintGL(QObject* context, std::unique_ptr<QOpenGLFunctions, void(*)(void*)> const &f);
-
-    virtual void UpdateScene(std::shared_ptr<Model> const &m, std::unique_ptr<QOpenGLFunctions, void(*)(void*)> const &f);
-
-    virtual MinMaxValue GetMinMax();
-
-    float CalcIdealSpacing();
-};
-
-
-class SceneLayer: public Layer
-{
-private:
-    std::unique_ptr<QOpenGLShaderProgram> program;
-    QOpenGLBuffer positionsBuffer;
-    QOpenGLBuffer valuesBuffer;
-    std::unique_ptr<std::vector<float>> positions;
-    std::unique_ptr<std::vector<float>> values;
-    GLuint textureID;
-    int lines;
-
-    void CreateColormap();
-
-public:
-    SceneLayer();
-
-    virtual void InitializeGL(QObject* context, std::unique_ptr<QOpenGLFunctions, void(*)(void*)> const &f);
-
-    virtual void PaintGL(QObject* context, std::unique_ptr<QOpenGLFunctions, void(*)(void*)> const &f);
-
-    virtual void UpdateScene(std::shared_ptr<Model> const &m, std::unique_ptr<QOpenGLFunctions, void(*)(void*)> const &f);
-
-    virtual MinMaxValue GetMinMax();
-};
-
 #endif //OPENPSTD_VIEWER2D_H
